Draft: move svg subfunctions to svgshapes.py module
