+++ src/Mod/Draft/CMakeLists.txt
@@ -98,6 +98,7 @@ SET(Draft_functions
     draftfunctions/scale.py
     draftfunctions/split.py
     draftfunctions/svg.py
+    draftfunctions/svgshapes.py
     draftfunctions/svgtext.py
     draftfunctions/upgrade.py
     draftfunctions/README.md
